remove -py3 related tests.  Move py3 specific code to it's own module
