--- Examples/python/import_packages/split_modules/builtin_nopy3/pkg1/Makefile
@@ -1,22 +1,0 @@
-TOP        = ../../../../..
-SWIGEXE    = $(TOP)/../swig
-SWIG_LIB_DIR = $(TOP)/../$(TOP_BUILDDIR_TO_TOP_SRCDIR)Lib
-SWIGOPT    = -builtin
-LIBS       =
-
-build:
-	$(MAKE) -f $(TOP)/Makefile SRCDIR='$(SRCDIR)' \
-        SWIGPYTHON='$$(SWIG) -python' \
-	SWIG_LIB_DIR='$(SWIG_LIB_DIR)' SWIGEXE='$(SWIGEXE)' \
-	SWIGOPT='$(SWIGOPT)' LIBS='$(LIBS)' TARGET='foo' \
-        INTERFACE='foo.i' python_cpp
-
-static:
-	$(MAKE) -f $(TOP)/Makefile SRCDIR='$(SRCDIR)' \
-	SWIG_LIB_DIR='$(SWIG_LIB_DIR)' SWIGEXE='$(SWIGEXE)' \
-	SWIGOPT='$(SWIGOPT)' LIBS='$(LIBS)' TARGET='foo' \
-        INTERFACE='foo.i' python_cpp_static
-
-clean:
-	$(MAKE) -f $(TOP)/Makefile SRCDIR='$(SRCDIR)' TARGET='foo' python_clean
-
