--- Examples/python/import_packages/split_modules/vanilla_nopy3/pkg1/foo.i
@@ -1,10 +1,0 @@
-%module(package="pkg1") foo
-%{
-static unsigned count(void)
-    {
-    return 3;
-    }
-%}
-
-unsigned count(void);
-
