--- Examples/python/import_packages/split_modules/vanilla_nopy3/runme.py
@@ -1,3 +1,0 @@
-import pkg1.foo
-
-assert(pkg1.foo.count() == 3)
