--- Examples/python/import_packages/split_modules/vanilla_split_nopy3/pkg1/__init__.py
@@ -1,1 +1,0 @@
-# killroy was here
